gametable object interface
move ball outside gametable
