+++ src/io/github/nrhoffmann/pong/gui/GameTable.java
@@ -4,17 +4,22 @@ import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.util.ArrayList;
 
 class GameTable extends JPanel{
     private Ball ball;
+    private ArrayList<Object> stuffOnTable = new ArrayList<>();
 
     GameTable() {
         setBackground(Color.DARK_GRAY);
         ball = new Ball();
+        stuffOnTable.add(ball);
 
         new Timer(5, e1 -> {
             Graphics g = getGraphics();
-            ball.tick();
+            for(Object o : stuffOnTable){
+                o.tick();
+            }
             paintComponent(g);
         }).start();
 
@@ -39,65 +44,8 @@ class GameTable extends JPanel{
         return ball;
     }
 
-    public class Ball {
-        private Color color;
-        private Point location;
-        private Vector vector;
-        private int size;
-        private int maxY;
-        private int minY;
-
-        public void setColor(Color color) {
-            this.color = color;
-        }
-
-        public Color getColor() {
-            return color;
-        }
-
-        public void setLocation(Point location) {
-            this.location = location;
-        }
-
-        public Point getLocation() {
-            return location;
-        }
-
-        public void setSize(int size) {
-            this.size = size;
-        }
-
-        public int getSize() {
-            return size;
-        }
-
-        public void setVector(Vector vector) {
-            this.vector = vector;
-        }
-
-        public Vector getVector() {
-            return vector;
-        }
-
-        void tick() {
-            int deltaX = (int) vector.getSpeedX();
-            int deltaY = (int) vector.getSpeedY();
-            if (location.y + deltaY < minY || location.y + deltaY > maxY) {
-                deltaY = Math.max(deltaY, minY - location.y); //todo this might be ugly, or it might simulate mass compression
-                deltaY = Math.min(deltaY, maxY - location.y);
-                vector.bounceY();
-            }
-            location.translate(deltaX, deltaY);
-        }
-
-        Ball() {
-            size = 500;
-            color = new Color(128, 128, 128);
-            location = new Point(GameTable.this.SIZE.width / 2 - size / 2, GameTable.this.SIZE.height / 2);
-            vector = new Vector();
-
-            maxY = GameTable.this.SIZE.height - size;
-            minY = 0;
-        }
+    interface Object {
+        abstract void tick();
     }
+
 }
